+++ packages/saas-ui-theme/src/base/components/nav-item.ts
@@ -18,6 +18,10 @@ const baseStyle: PartsStyleFunction<typeof parts> = (props) => {
   return {
     item: {
       my: '2px',
+      color: 'gray.900',
+      _dark: {
+        color: 'whiteAlpha.900',
+      },
     },
     link: {
       display: 'flex',
@@ -27,10 +31,6 @@ const baseStyle: PartsStyleFunction<typeof parts> = (props) => {
       textDecoration: 'none',
       transitionProperty: 'common',
       transitionDuration: 'normal',
-      color: 'gray.900',
-      _dark: {
-        color: 'whiteAlpha.900',
-      },
       _hover: {
         textDecoration: 'none',
       },
@@ -50,14 +50,15 @@ const baseStyle: PartsStyleFunction<typeof parts> = (props) => {
         justifyContent: 'center',
       },
     },
+    label: {
+      '[data-condensed] &': {
+        display: 'none',
+      },
+    },
     icon: {
       display: 'flex',
       transitionProperty: 'common',
       transitionDuration: 'normal',
-      color: 'gray.900',
-      _dark: {
-        color: 'whiteAlpha.900',
-      },
       '[data-condensed] &': {
         me: 0,
         alignItems: 'center',
@@ -106,11 +107,8 @@ const variantSubtle: PartsStyleFunction<typeof parts> = (props) => {
       _active: {
         bg: activeBg,
         fontWeight: 'semibold',
+        color: mode(`${c}.600`, `${c}.100`)(props),
       },
-      color,
-    },
-    icon: {
-      color,
     },
   }
 }