+++ packages/saas-ui-theme/src/saas-ui/components/tooltip.ts
@@ -1,18 +1,22 @@
-import { mode, cssVar, SystemStyleFunction } from '@chakra-ui/theme-tools'
+import { cssVar, SystemStyleFunction } from '@chakra-ui/theme-tools'
 
 const $bg = cssVar('tooltip-bg')
+const $fg = cssVar('tooltip-fg')
 const $arrowBg = cssVar('popper-arrow-bg')
 
 const baseStyle: SystemStyleFunction = (props) => {
-  const bg = mode('white', 'gray.700')(props)
   return {
     display: 'flex',
-    [$bg.variable]: `colors.${bg}`,
+    [$bg.variable]: 'colors.white',
+    [$fg.variable]: 'colors.blackAlpha.900',
+    _dark: {
+      [$bg.variable]: 'colors.gray.700',
+      [$fg.variable]: 'colors.whiteAlpha.900',
+    },
     px: '8px',
     py: '2px',
     bg: [$bg.reference],
     [$arrowBg.variable]: [$bg.reference],
-    color: mode('blackAlpha.900', 'whiteAlpha.900')(props),
     borderRadius: 'sm',
     fontWeight: 'medium',
     fontSize: 'xs',