+++ packages/saas-ui-app-shell/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @saas-ui/app-shell
 
+## 1.0.0
+
+### Major Changes
+
+- e670761: AppShell moved out of beta. 🥳
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
 ## 0.2.0
 
 ### Minor Changes