+++ packages/saas-ui-clerk/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @saas-ui/clerk
 
+## 1.2.5
+
+### Patch Changes
+
+- @saas-ui/auth@1.6.1
+
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies [d92f516]
+  - @saas-ui/auth@1.6.0
+
 ## 1.2.3
 
 ### Patch Changes