+++ packages/saas-ui-menu/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @saas-ui/menu
 
+## 1.3.1
+
+### Patch Changes
+
+- @saas-ui/provider@1.1.7
+
+## 1.3.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [d92f516]
+  - @saas-ui/button@1.3.0
+  - @saas-ui/provider@1.1.6
+
 ## 1.2.2
 
 ### Patch Changes