+++ packages/saas-ui-sidebar/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @saas-ui/sidebar
 
+## 0.5.5
+
+### Patch Changes
+
+- Fixed issue where SSR on mobile devices would not render properly.
+
+## 0.5.4
+
+### Patch Changes
+
+- 7cf5223: Improved toggle button rendering and behavior with ssr / mobile screens.
+- Updated dependencies [7cf5223]
+  - @saas-ui/react-utils@1.1.2
+  - @saas-ui/provider@1.1.7
+
+## 0.5.3
+
+### Patch Changes
+
+- Only set tooltip label if children is a string.
+
+## 0.5.2
+
+### Patch Changes
+
+- Fix SSR variant value.
+
+## 0.5.1
+
+### Patch Changes
+
+- ddb0019: Fix React hydration error with condensed variant
+
+## 0.5.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [d92f516]
+  - @saas-ui/collapse@1.3.0
+  - @saas-ui/provider@1.1.6
+
 ## 0.4.0
 
 ### Minor Changes