+++ packages/saas-ui-theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @saas-ui/theme
 
+## 1.6.1
+
+### Patch Changes
+
+- 7cf5223: Improved Sidebar theme added toggle and toggleWrapper parts.
+
+## 1.6.0
+
+### Minor Changes
+
+- d92f516: Updated to Chakra UI 2.4.1
+
 ## 1.5.0
 
 ### Minor Changes